+++ core/src/main/java/timeboard/core/model/Project.java
@@ -26,6 +26,8 @@ package timeboard.core.model;
  * #L%
  */
 
+import timeboard.core.model.converters.JSONToProjectAttributsConverter;
+
 import javax.persistence.*;
 import java.io.Serializable;
 import java.util.*;
