+++ core/src/main/java/timeboard/core/model/converters/JSONToStringListConverter.java
@@ -1,4 +1,4 @@
-package timeboard.core.model;
+package timeboard.core.model.converters;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
@@ -7,11 +7,10 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import javax.persistence.AttributeConverter;
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.HashMap;
+
 import java.util.List;
 
-public class StringListToStringConverter
+public class JSONToStringListConverter
        implements AttributeConverter<List<String>, String> {
     private static final String DELIMITER = "|";
